External: Allow references to paths relative to output path
Sometimes it is necessary for a step to produce more than one output file. e.g.
one might need a command line of the form,

    $executable --out1 $output/out1 --out2 $output/out2

Where $output is the output path in the content store.

Here we add support for this by allowing ParamField to capture paths relative to
the output path. To avoid the pain of having to quantify over the File/Dir
phanton of Path, we introduce two constructors: ParamOutFile and ParamOutDir.